+++ example/Smartflow.BussinessService/Smartflow.BussinessService.csproj
@@ -66,7 +66,9 @@
       <Name>Dapper</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Compile Include="WorkflowService\BaseWorkflowService.Process.cs" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
